backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.1.0
pytest==7.3.1
pytest-django==4.5.2
